workshop/local/requirements.txt
cloudml-hypertune==0.1.0.dev6
fire==0.3.1
google-cloud-bigquery==2.5.0
google-cloud-bigquery-storage==2.3.0
pandas==1.1.4
pyarrow==2.0.0
scikit-learn==0.23.2
xgboost==1.2.1

google-api-python-client==1.12.8
google-cloud-storage==1.33.0
gcsfs==0.7.1
